% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{countMatrix}
\alias{countMatrix}
\title{PWM number comparison}
\usage{
countMatrix(target_mot_posi, ctrl_mot_posi, seq_range, windowSize = 100,
  slide = 50)
}
\arguments{
\item{target_mot_posi}{PWM positions of target regions}

\item{ctrl_mot_posi}{PWM positions of randomely selected regions}

\item{seq_range}{range from probe position to be analyze}

\item{windowSize}{window size}

\item{slide}{length of window sliding}
}
\value{
a matrix containg number of PWM at demethylated regions, number of PWM at random regions, TF, and adjust p-value
}
\description{
Count PWM in a window and compute fold-change(FC) and poisson p-value.
}
\keyword{PWM,}
\keyword{count}
\keyword{fold-cange,}
\keyword{of}
\keyword{p-value}
