% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DmpId.R
\name{DmpId}
\alias{DmpId}
\title{Identificaion of differentially methylated probes}
\usage{
DmpId(selDataMatrix, ControlColnum, TreatmentColnum, p.cutoff = 0.05,
  cutoff = 2, MethylDemethyl = "Demethyl")
}
\arguments{
\item{selDataMatrix}{normalized M-value matrix}

\item{ControlColnum}{column number for Control sample(s)}

\item{TreatmentColnum}{column number for target sample(s)}

\item{p.cutoff}{cut-off for Welch's t-test}

\item{cutoff}{cut-off for delta-M value}

\item{MethylDemethyl}{direction of Ddifferentially methylated regions (methylated or demethylated)}
}
\value{
IDs of differentially methylated probes
}
\description{
Compare two data (or data sets) conpute differentially methylated probes.
For single data comparison, delta-M value is used.
For data set comparison, both Welch's t-test and delta-M value are used.
}
\keyword{differentially}
\keyword{methylated}
\keyword{probe}
