% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichScore.R
\name{enrichScore}
\alias{enrichScore}
\title{Computation of a enrichment score}
\usage{
enrichScore(x, y, z, nProbeList)
}
\arguments{
\item{x}{A position from the DMP}

\item{y}{PWM distribution at a DMP region}

\item{z}{motif distribution of s randomly selected probe region}

\item{nProbeList}{number of DMP}
}
\value{
a enrichment score
}
\description{
Mixture distribution of karnel densities per region. 
karnel = gaussian
}
\keyword{PWM,}
\keyword{density}
\keyword{distribution}
\keyword{enrichmen}
\keyword{karnel}
\keyword{mixture}
\keyword{tscore,}
