% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichScoreDist.R
\name{enrichScoreDist}
\alias{enrichScoreDist}
\title{Enrichment score}
\usage{
enrichScoreDist(target_mot_posi, ctrl_mot_posi, seq_range,
  motif_name = "motif", nDMP_IDs, outname = "enrichment_score",
  plot_draw = TRUE)
}
\arguments{
\item{target_mot_posi}{PWM positions from the DMPs}

\item{ctrl_mot_posi}{PWM positions from the probe of random rprobes}

\item{seq_range}{range to be analized from the DMPs}

\item{motif_name}{PWM ID}

\item{nDMP_IDs}{number of DMPs}

\item{outname}{ID for output file}

\item{plot_draw}{ligical. if TRUE, output a plot of enrichment score.}
}
\value{
enrucment score
}
\description{
Plot of enrichment score and return enrichment scores
The enrichment score is conputed based on modified mixture distribution of
}
\keyword{PWM,}
\keyword{distribution}
\keyword{enrichment}
\keyword{mixture}
\keyword{score,}
