% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichTest.R
\name{enrichTest}
\alias{enrichTest}
\title{Significant Calling and summary output}
\usage{
enrichTest(significant_ranges, motif_counts_matrix, seq_range,
  enrichment_scores, windowSize)
}
\arguments{
\item{significant_ranges}{generated by SigRange}

\item{motif_counts_matrix}{generated by countMatrix}

\item{seq_range}{Analyzed range from DMP}

\item{enrichment_scores}{enrichScoreDist}

\item{windowSize}{windowsize used for p-value coumutation etc... (should be same as windowSize parameter of countMatrix)}
}
\value{
summary of analysis and signicifant call
}
\description{
Judge whether PWM is significantly overrepresented at the DMP regions.
The Judgement is made by Poisson p-value at DMP position and enrichment score at DMP position which is more than Q3+IQR*3 of enrichment score of analyzed range.
}
\keyword{call}
\keyword{significant}
\keyword{summary,}
