% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfiniumDiffMetMotR.R
\docType{data}
\name{epic.controls}
\alias{epic.controls}
\title{EPIC control database}
\format{635 abs. of 4 variables}
\usage{
data(epic.controls)
}
\description{
EPIC control dataset from wateRmelon
}
\keyword{EPIC}
\keyword{wateRmelon}
