% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfiniumDiffMetMotR.R
\docType{data}
\name{EPICanno}
\alias{EPICanno}
\title{Annotation of InfiniumMethylationEPIC beadsChip}
\format{A data frame with 867531 rows and 6 variables}
\usage{
data(EPICanno)

data(EPICanno)
}
\description{
a list containing the annotation of each probe. The variables are as follows:

a list containing the annotation of each probe. The variables are as follows:
}
\details{
\itemize{
\item IlmnID. probe ID
\item Genome_Build. version of genome
\item CHR. chromosome
\item MAPINFO. coordinate of genome
UCSC_RefGene_Group. Annotation against gene
Relation_to_UCSC_CpG_Island. Annotation against CpG Island
}

\itemize{
\item IlmnID. probe ID
\item Genome_Build. version of genome
\item CHR. chromosome
\item MAPINFO. coordinate of genome
UCSC_RefGene_Group. Annotation against gene
Relation_to_UCSC_CpG_Island. Annotation against CpG Island
}
}
\keyword{850}
\keyword{EPIC}
\keyword{Infinium}
\keyword{annotation}
\keyword{datasets}
