% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCChangePlot.R
\name{FCChangePlot}
\alias{FCChangePlot}
\title{Plot a Fold-Change plot}
\usage{
FCChangePlot(motif_name = "motif", motif_counts_matrix)
}
\arguments{
\item{motif_name}{PWM ID}

\item{motif_counts_matrix}{motif counts matrix output from countMatrix}
}
\description{
Vidualization of FC
}
\keyword{fold-cange,}
\keyword{plot}
