% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer2PWMfile.R
\name{homer2PWMfile}
\alias{homer2PWMfile}
\title{Homer PWM to PWM for InfiniumDiffMetMotR}
\usage{
homer2PWMfile(file)
}
\arguments{
\item{file}{homer PWM file}
}
\value{
PWM list for InfiniumDiffMetMotR
}
\description{
Homer PWM to PWM for InfiniumDiffMetMotR
}
\keyword{InfiniumDiffMetMotR,}
\keyword{PWM}
\keyword{homer,}
