% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfiniumDiffMetMotR.R
\docType{data}
\name{IMAGE_PWMlist}
\alias{IMAGE_PWMlist}
\title{IMAGE PWM database}
\format{A list with 1611 elements}
\usage{
data(IMAGE_PWMlist)
}
\description{
a list containing the annotation of each probe.
}
\keyword{IMAGE}
\keyword{PWM}
\keyword{datasets}
\keyword{list}
