% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumiMethyNorm.R
\name{lumiMethyNorm}
\alias{lumiMethyNorm}
\title{Infinium Normalization & M-value computation}
\usage{
lumiMethyNorm(fileName = "TableControl.txt", idatpath = getwd(),
  inputtype, sample_names = FALSE)
}
\arguments{
\item{fileName}{unnourmalized intensity data}

\item{idatpath}{directroy of idata files}

\item{inputtype}{type of input. "signal": a row signal txt file generated by genomeStudio or "idat": idat files}

\item{sample_names}{a vetor of sample names}
}
\value{
normalized M-value
}
\description{
Normalize the Infinium signal intensitty data and compute M-value.
Input data should be unnormalized data containing;
TargetID: Illumina ID
Sample1.Signal_A: Signal of Unmethylated probe of sample 1
Sample1.Signal_B: Signal of Methylated probe of sample 1
Sample1.Detection Pval: detection P value of sample 1
(sample_names: vector of sample names)
}
\keyword{M-value}
\keyword{Normalization,}
\keyword{lumi,}
