% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfiniumDiffMetMotR.R
\docType{data}
\name{Methyl450anno}
\alias{Methyl450anno}
\title{Annotation of InfiniumMethylation450 beadsChip}
\format{A data frame with 486428 rows and 6 variables}
\usage{
data(Methyl450anno)
}
\description{
a dataset containing the annotation of each probe. The variables are as follows:
}
\details{
\itemize{
\item IlmnID. probe ID
\item Genome_Build. version of genome
\item CHR. chromosome
\item MAPINFO. coordinate of genome
UCSC_RefGene_Group. Annotation against gene
Relation_to_UCSC_CpG_Island. Annotation against CpG Island
}
}
\keyword{450}
\keyword{Infinium}
\keyword{Methyl450}
\keyword{annotation}
\keyword{datasets}
