% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motHist.R
\name{motHist}
\alias{motHist}
\title{Draw a Histgram}
\usage{
motHist(target_mot_posi, ctrl_mot_posi, seq_range, motif_name = "motif")
}
\arguments{
\item{target_mot_posi}{PWM positions from the DMPs}

\item{ctrl_mot_posi}{PWM positions from the probe of random rprobes}

\item{seq_range}{range to be analized from the DMPs}

\item{motif_name}{PWM ID}
}
\description{
Vidualize the PWM distribution for target and randomly selected regions by histgram.
}
\keyword{PWM}
\keyword{distribution}
\keyword{histgram,}
