% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifClassCount.R
\name{motifClassCount}
\alias{motifClassCount}
\title{PWM Count}
\usage{
motifClassCount(motifPosi, windowSize, slide, seq_range)
}
\arguments{
\item{motifPosi}{relative position of identified PWM from probe}

\item{windowSize}{window size}

\item{slide}{length of window sliding}

\item{seq_range}{range from probe position to be analyze}
}
\value{
a data frame of class and number of PWM.
}
\description{
cunting PWMs in a window, sliding the window.
}
\keyword{PWM}
\keyword{count}
\keyword{of}
