% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifDistFromCpG.R
\name{motifDistFromCpG}
\alias{motifDistFromCpG}
\title{PWM serch with singel motif and single sequence}
\usage{
motifDistFromCpG(pfm, DNAseq, half_length)
}
\arguments{
\item{pfm}{A PWM}

\item{DNAseq}{A sequence}

\item{half_length}{half of sequence length}
}
\value{
mached PWM positions (distances from probe)
}
\description{
search a pfm and output the distance from center
The default minimam score = 90%
}
\keyword{PWM,}
\keyword{motif}
\keyword{seaech}
