% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifDistMultiSeqs.R
\name{motifDistMultiSeqs}
\alias{motifDistMultiSeqs}
\title{Parallel of motifDistFromCpG with a PWM}
\usage{
motifDistMultiSeqs(motif, fasta2)
}
\arguments{
\item{motif}{A PWM}

\item{fasta2}{A DNAStringSet sequnce}
}
\value{
A list of identified motif positions in a region.
}
\description{
Run motifDistFromCpG, a motif search function, for one PWM in parallel for multiple sequences with multi-CPU.
Default CPU number is 4
}
\keyword{PWM,}
\keyword{motif}
\keyword{parallel}
\keyword{seaech,}
