% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotScr.R
\name{MotScr}
\alias{MotScr}
\title{PWM overrepresentation analysis pipline for illumina methylation array}
\usage{
MotScr(infile = "sel_processed_Mval.txt", motifDBList, cutoff = 2,
  p.cutoff = 0.001, outname = "screening_result", ControlColnum,
  TreatmentColnum, MethylDemethyl = "Demethyl", version = "850",
  sampling = FALSE)
}
\arguments{
\item{infile}{input file: normalized M-value matrix}

\item{motifDBList}{PWM data list}

\item{cutoff}{cutoff for delta-M-value}

\item{p.cutoff}{cutoff for Welch's t-test}

\item{outname}{output files ID}

\item{ControlColnum}{column number for Control sample(s)}

\item{TreatmentColnum}{column number for target sample(s)}

\item{MethylDemethyl}{direction of Ddifferentially methylated regions (methylated or demethylated)}

\item{version}{version of Infinium Methjylation array (450 or 850(orEPIC))}

\item{sampling}{If sampling number is indicated and DMP is more than sampling number, analysis is ran with randomly selected DMP. If FALSE, all DMPs are analyzed.}
}
\value{
analysis summary files and plots, Rdata
}
\description{
pipline to analyze enrichment of given motif PWMs for differentially methylated region based on Illumina infinium methylation array data.
}
\keyword{PWM,}
\keyword{analysis}
\keyword{overrepresentation}
