% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisTestPlot.R
\name{PoisTestPlot}
\alias{PoisTestPlot}
\title{Plot a p-value plot}
\usage{
PoisTestPlot(motif_counts_matrix, motif_name = "motif", seq_range)
}
\arguments{
\item{motif_counts_matrix}{motif counts matrix output from countMatrix}

\item{motif_name}{PWM ID}

\item{seq_range}{range to be analyzed}
}
\description{
Vidualization of the Poisson test p-value.
The value is -10logP
red dots represent p <= 0.00001
}
\keyword{Poisson}
\keyword{p-value,}
\keyword{plot}
\keyword{test,}
