% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeID2position.R
\name{probeID2position}
\alias{probeID2position}
\title{conversion of probe ID to genomic corrdinate}
\usage{
probeID2position(probe_IDs, anno_info = InfiniumDiffMetMotR::EPICanno)
}
\arguments{
\item{probe_IDs}{probe IDs of methylation array}

\item{anno_info}{annotation data base (Methyl450anno for Methyl450 array and EPICanno for EPIC array)}
}
\value{
genomic coordinates
}
\description{
#conversion from illumina methylation array ID to genomic corrdinate (hg19)
}
\keyword{ID,}
\keyword{coordinate,}
\keyword{genoic}
\keyword{genomic}
\keyword{position}
\keyword{probe}
