% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{betas}
\alias{methylated}
\alias{unmethylated}
\alias{pval.detect<-}
\alias{reverseComplement}
\alias{width}
\alias{matchPWM}
\alias{start}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Biostrings}{\code{\link[Biostrings]{reverseComplement}}, \code{\link[Biostrings]{width}}, \code{\link[Biostrings]{matchPWM}}, \code{\link[Biostrings]{start}}}

  \item{methylumi}{\code{\link[methylumi]{betas}}, \code{\link[methylumi]{methylated}}, \code{\link[methylumi]{unmethylated}}, \code{\link[methylumi]{pval.detect<-}}}
}}

