% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selRawMetSamples.R
\name{selRawMetSamples}
\alias{selRawMetSamples}
\title{data selection for raw Infinium signal data}
\usage{
selRawMetSamples(TableControl = "TableControl.txt", sampleIDs)
}
\arguments{
\item{TableControl}{unnormalized no background corrction raw methylation array data}

\item{sampleIDs}{a vector of sample IDs to be selected}
}
\value{
selected unnormalized no background corrction raw methylation array data
}
\description{
This function extracts necessary data of targets from the infinium array raw data table.
}
\keyword{Infinium,}
\keyword{sampleID}
