% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRange.R
\name{SigRange}
\alias{SigRange}
\title{Extraction of significantly enriched ranges}
\usage{
SigRange(motif_counts_matrix, cutoff = 1e-05, windowSize)
}
\arguments{
\item{motif_counts_matrix}{motif counts matrix output from countMatrix}

\item{cutoff}{cutoff p-value}

\item{windowSize}{windowsize used for p-value coumutation (should be same as windowSize parameter of countMatrix)}
}
\value{
range of the regions where PWM is significantly enrihed.
}
\description{
Extraction of significantly enriched ranges based on poisson test p-value. The positions where PWM is significantly enriched are marged as a region.
}
\keyword{Poisson}
\keyword{p-value,}
\keyword{region}
\keyword{test,}
