% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSeqMotDist.R
\name{splitSeqMotDist}
\alias{splitSeqMotDist}
\title{PWM serch for splitted sequence files}
\usage{
splitSeqMotDist(filenames, motif_list)
}
\arguments{
\item{filenames}{list of splitted files}

\item{motif_list}{PWM list}
}
\value{
relative positions of PWM from the probes
}
\description{
PWM serch for all solitted sequence files
#Splitted seqs are sequentially read and serched the given PWM
}
\keyword{PWM,}
\keyword{motif}
\keyword{seaech}
