% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratSampling.R
\name{stratSampling}
\alias{stratSampling}
\title{Stratified sampling}
\usage{
stratSampling(target_IDs, anno_info = InfiniumDiffMetMotR::EPICanno)
}
\arguments{
\item{target_IDs}{methylation array IDs of DMP}

\item{anno_info}{annotation infromation of probe}
}
\value{
probe IDs
}
\description{
stratified sampling of same number of target probes based on annotation categolies of target probes(TSS200, CGI, CpG shore, CpG shelf, and other region)
}
\keyword{random}
\keyword{sampling,}
\keyword{stratified}
