% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSplitSeq.R
\name{writeSplitSeq}
\alias{writeSplitSeq}
\title{Write sequences to multiple output files}
\usage{
writeSplitSeq(seqs, split_num = 2500, tempDir = "Transient.out",
  output_file = "transient.out")
}
\arguments{
\item{seqs}{DNAStringSet sequences}

\item{split_num}{nuber of sequences per file}

\item{tempDir}{directory to be stored the sequences}

\item{output_file}{file ID for output files}
}
\value{
output filenames
}
\description{
To reduce the memory occupancy, the DNAStringSet sequences are output to files, splitted into indicate numbers.
}
\keyword{DNAStringSet}
\keyword{sequences}
\keyword{split}
