% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdat2M.R
\name{getIdat2M}
\alias{getIdat2M}
\title{GetGEO idat file & M-value conversion}
\usage{
getIdat2M(GEOid, sampleNames = FALSE)
}
\arguments{
\item{GEOid}{GEO ID}

\item{sampleNames}{a vector of sampeNamse. If "FALSE", sample namses use GEO sample titles.}
}
\value{
Normalized M-value text files
}
\description{
Downloads idat files based given GEO ID, normalzies, and computes M-values.
}
\keyword{GEO}
\keyword{M-value}
\keyword{Normalization}
\keyword{idat}
